requirements.txt
jupyter
matplotlib
numpy
scikit-learn
pandas
pillow